"Capture page info" tweak
Made it run even on `PARTIAL` state.

This step grabs info about the page from the browser and also tries to grab the favicon if missing (it is not pulled automatically in headless mode).

Replaced fetch call for doing that with `curl` + `ScoopProxy` as a way to enforce `PARTIAL` state and `blocklist`.

Bonus: the favicon is no longer a generated exchange in that case.